requirements.txt
alabaster==0.7.12
asgiref==3.2.3
Babel==2.8.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
Django==2.2.7
django-cors-headers==3.2.1
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
docutils==0.16
idna==2.8
imagesize==1.2.0
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.1
pycparser==2.19
Pygments==2.5.2
pyparsing==2.4.6
pytz==2019.3
requests==2.22.0
simplejson==3.17.0
six==1.14.0
snowballstemmer==2.0.0
Sphinx==2.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
